cdn. -> c. (#185)
* cdn. -> c.

I had a sinking suspicion that Cloudflare was doing something to cdn.chrisshort.net after the AWS bill incident. All monitoring tools are getting 403s where all humans aren't. cache is still an option but, I'd rather use something shorter so c.chrisshort.net it is (unless that's screwed up by Cloudflare too, in which case I will move it over to AWS) and not try to CDN on Cloudflare anymore.

* Create apple-touch-icon.png

Integrity 404s on this. Not sure if it's needed but what's one more png amongst friends.

Former-commit-id: b33ccd37bd7ff5c79e34bd4aa66bdee50e3e11a9
Former-commit-id: 2fa2a5b46828f415b17da0d31fb04fa56bb6cba0